% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_dl.R
\name{download_station_data}
\alias{download_station_data}
\title{Download data for a single station}
\usage{
download_station_data(stn_id, collection, start_date, end_date, verbose)
}
\description{
Download data for a single station
}
\keyword{internal}
