% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_variables}
\alias{get_variables}
\title{Get available weather variables for an interval}
\usage{
get_variables(interval = c("hour", "day", "month"))
}
\arguments{
\item{interval}{Character. Time interval: "hour", "day", or "month".}
}
\value{
A character vector of available variables
}
\description{
Returns a list of available weather variables for the specified interval.
}
\examples{
\dontrun{
get_variables("day")
}
}
