% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{STATIONS_CACHE_FILE}
\alias{STATIONS_CACHE_FILE}
\title{Cache file for stations data}
\format{
An object of class \code{character} of length 1.
}
\usage{
STATIONS_CACHE_FILE
}
\description{
Cache file for stations data
}
\keyword{datasets}
