% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stations_meta}
\alias{stations_meta}
\title{Get stations metadata information}
\usage{
stations_meta()
}
\value{
A list with metadata about the stations data
}
\description{
Returns metadata about the cached stations data, including when it was last
downloaded and the number of stations available.
}
\examples{
\dontrun{
stations_meta()
}
}
