% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weathercan2-package.R
\docType{package}
\name{weathercan2}
\alias{weathercan2-package}
\alias{weathercan2}
\title{weathercan2: Download and Format Weather Data from ECCC}
\description{
An R package for downloading historical weather data from Environment and Climate
Change Canada (ECCC) using the MSC GeoMet API. This package provides functions
to search for weather stations and download daily, hourly, and monthly weather
observations.
}
\section{Main Functions}{

\itemize{
\item \code{\link{stations}} - Get all climate stations
\item \code{\link{stations_search}} - Search for stations by name, location, or ID
\item \code{\link{weather_dl}} - Download weather data for stations
\item \code{\link{normals_dl}} - Download climate normals
}
}

\section{API}{

This package uses the MSC GeoMet API (https://api.weather.gc.ca/) which provides
access to Environment and Climate Change Canada weather and climate data.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yourusername/weathercan2}
  \item Report bugs at \url{https://github.com/yourusername/weathercan2/issues}
}

}
\author{
\strong{Maintainer}: Your Name \email{your.email@example.com}

}
